Bdcalling chatbot/requirements.txt
openai>=1.3.0
streamlit
python-dotenv
